Linear  Regression
In this project, you will be working with a dataset that contains information about the population and profit of a food business in different cities. 
Your goal is to implement a linear regression model from scratch using Python and NumPy to predict the profit based on the population of a city.